kownledgeBase_response_from_crewAI/requirements.txt
Flask
python-dotenv
crewai
crewai-tools

langchain-huggingface
